updated start logic for multiplayer, added sub-goals + convoluted logic
Changed settings creation into a large dictionary (gets .Key and .Value as the setting's properties).
Populates a sub-goal dictionary at the same time. This is necessary so double splits don't occur (if you select the final sub-goal in your gameplay, it would split for the sub-goal and the overall goal). To circumvent duplicate splits (if a sub-goal is 'unawarded' (i.e. an item falls off the basket during the 'go shopping' goal in high street), it should not split again upon re-awarding), a duplicate dictionary is created. The original dictionary has sub-goals removed as they happen. It must be re-populated when a new run starts.
Unsure if this logic can be shortened, if you have the time, please take a look.

Starting for multiplayer now makes the honk variable jump from 16 to 18, 16 to 17 still remains for single player.