added splitting upon getting the crowns, moved variable resetting
Moved resetting the variables to their defaults from the start block to a logic that returns when the timer is reset (manually or not). This is to avoid any potential lag when the dictionary re-populates before starting the timer. This is not put into the reset { } block to accomodate for potential occurances where the player resets manually.

Added splits for both the normal and 2nd player crowns.